add logs

+++ simple-calculator/src/test/java/com/tuanmhoang/calculator/CalculatorTest.java
@@ -1,5 +1,6 @@
 package com.tuanmhoang.calculator;
 
+import java.math.BigInteger;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -8,17 +9,17 @@ public class CalculatorTest {
 
     @Test
     public void givenTwoNumber_whenSumIsCalled_thenReturnResult() {
-        assertEquals(5, Calculator.sum(2, 3));
+        assertEquals(BigInteger.valueOf(5), Calculator.sum(2, 3));
     }
 
     @Test
     public void givenTwoNumber_whenMinusIsCalled_thenReturnResult() {
-        assertEquals(-1, Calculator.minus(2, 3));
+        assertEquals(BigInteger.valueOf(-1), Calculator.minus(2, 3));
     }
 
     @Test
     public void givenTwoNumber_whenMultiplyIsCalled_thenReturnResult() {
-        assertEquals(6, Calculator.multiply(2, 3));
+        assertEquals(BigInteger.valueOf(6), Calculator.multiply(2, 3));
     }
 
 }
